Reset adjusted altitude if ground altitude no longer applicable
This fixes a major issue with other aircraft showing at the wrong altitude. Below FL180, xPilot probes the terrain elevation so that it can smoothly apply a terrain offset altitude (i.e. to smoothly show aircraft landing or taking off, without sudden altitude jumps). When the aircraft climbs above FL180, the adjusted altitude isn't reset, so xPilot is setting the altitude of the aircraft to a lower altitude based on terrain probing (presumably, the aircraft is stuck at FL180).