Additional tweaks to Windows installer
This should resolve the issue with the installer freezing for some people. If the user installed xPilot in a folder that contained other non-xPilot files, the installer would try to add all non-xPilot files also inside the folder to the uninstaller; if there were lots of files, the installer would essentially freeze. Now the installer doesn't explicitly include all files in the install directory. The uninstaller will now only remove the files that were installed by the installer (read: only xPilot application files).